rm level of indirection
